requirements.txt
Flask==2.0.3
flask-socketio==5.3.0
opencv-python==4.5.5.64
opencv-python-headless==4.5.5.64
mysql-connector-python==8.0.26
ultralytics==8.0.19
numpy==1.21.2
Pillow==8.4.0
PyYAML==5.4.1
requests==2.26.0
